Move dom manipulation into index.js
